moved autoencoder models to their own folder
